优化http框架的逻辑

+++ edap-common/src/main/java/io/edap/util/UnsafeUtil.java
@@ -134,7 +134,7 @@ public class UnsafeUtil {
         return UNSAFE.getByte(value, offset);
     }
 
-    public static int readByte(long address) {
+    public static byte readByte(long address) {
         return UNSAFE.getByte(address);
     }
